Add swagger, refactor code, fix bags

+++ UserMicroService/UserService.Presentation/Controllers/UsersController.cs
@@ -1,23 +1,28 @@
 ﻿using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Service.Contract;
 using Shared.DataTransferObjects.UserDto;
+using System.Security.Claims;
 namespace UserService.Presentation.Controllers
 {
     [Route("api/users")]
     [ApiController]
-    
-    public class UsersController(IServiceManager _serviceManager,  IValidator<UserForUpdateDto> _updateValidator) : ControllerBase
+    public class UsersController(IServiceManager _serviceManager, IValidator<UserForUpdateDto> _updateValidator) : ControllerBase
     {
+        [Authorize(Roles = "User,Admin")]
         [HttpGet("{id:guid}", Name = "UserById")]
         public async Task<IActionResult> GetUserById(Guid id)
         {
-            Console.WriteLine("rt");
+            if (!await CanAccessUserAsync(User, id))
+                return Forbid();
+            
             var user = await _serviceManager.UserService.GetUserById(id, trackChanges: false);
 
             return Ok(user);
         }
 
+        [Authorize(Roles = "Admin")]
         [HttpGet("collection/({ids})", Name = "UserCollection")]
         public async Task<IActionResult> GetUsersCollections(IEnumerable<Guid> ids)
         {
@@ -26,23 +31,44 @@ namespace UserService.Presentation.Controllers
             return Ok(users);
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpDelete("{id:guid}")]
         public async Task<IActionResult> DeleteUser(Guid id)
         {
+            if (!await CanAccessUserAsync(User, id))
+                return Forbid();
+
             await _serviceManager.UserService.DeleteUser(id, trackChanges: false);
 
             return NoContent();
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPut("{id:guid}")]
         public async Task<IActionResult> UpdateUser(Guid id, [FromBody] UserForUpdateDto userForUpdate)
         {
             _updateValidator.ValidateAndThrow(userForUpdate);
+
+            if (!await CanAccessUserAsync(User, id))
+                return Forbid();
 
             await _serviceManager.UserService.UpdateUser(id, userForUpdate, trackChanges: true);
 
             return NoContent();
         }
 
+
+        private Task<bool> CanAccessUserAsync(ClaimsPrincipal user, Guid targetUserId)
+        {
+            var currentUserId = user.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isUser = user.IsInRole("User");
+
+            if (isUser && targetUserId.ToString() != currentUserId)
+            {
+                return Task.FromResult(false);
+            }
+
+            return Task.FromResult(true);
+        }
     }
 }
